[Group-Service] 공고 API 오류 수정

See merge request s09-final/S09P31A603!236

+++ tunemate-group-service/src/main/java/kr/co/tunemate/tunemategroupservice/service/GroupParticipationServiceImpl.java
@@ -1,17 +1,20 @@
 package kr.co.tunemate.tunemategroupservice.service;
 
+import kr.co.tunemate.tunemategroupservice.dto.layertolayer.GroupDto;
 import kr.co.tunemate.tunemategroupservice.dto.layertolayer.GroupParticipationDto;
 import kr.co.tunemate.tunemategroupservice.entity.GroupParticipation;
 import kr.co.tunemate.tunemategroupservice.exception.BaseException;
 import kr.co.tunemate.tunemategroupservice.exception.code.GroupErrorCode;
 import kr.co.tunemate.tunemategroupservice.repository.GroupParticipationRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class GroupParticipationServiceImpl implements GroupParticipationService {
@@ -26,8 +29,17 @@ public class GroupParticipationServiceImpl implements GroupParticipationService 
      */
     @Override
     public List<GroupParticipationDto> findByUserId(String userId) {
+        log.info("사용자(userId: {})가 참여중인 공고 목록을 조회합니다.", userId);
+
         return groupParticipationRepository.findAllByUserId(userId).stream()
-                .map(groupParticipation -> modelMapper.map(groupParticipation, GroupParticipationDto.class)).toList();
+                .map(groupParticipation -> {
+                    GroupParticipationDto groupParticipationDto = modelMapper.map(groupParticipation, GroupParticipationDto.class);
+                    GroupDto groupDto = modelMapper.map(groupParticipation.getGroup(), GroupDto.class);
+
+                    groupParticipationDto.setGroupDto(groupDto);
+
+                    return groupParticipationDto;
+                }).toList();
     }
 
     /**
@@ -38,10 +50,16 @@ public class GroupParticipationServiceImpl implements GroupParticipationService 
     @Transactional
     @Override
     public void deleteByGroupParticipationId(String userId, String groupParticipationId) {
+        log.info("사용자(userId: {})가 참여중인 공고(groupParticipationId: {})를 탈퇴합니다.", userId, groupParticipationId);
+
         GroupParticipation groupParticipation = groupParticipationRepository.findByGroupParticipationId(groupParticipationId).orElseThrow(() -> new BaseException("존재하지 않는 공고참여입니다.", GroupErrorCode.NO_SUCH_ITEM_EXCEPTION.getHttpStatus()));
 
         if (!groupParticipation.getUserId().equals(userId)) {
             throw new BaseException("자신의 공고참여에 대해서만 탈퇴를 요청할 수 있습니다.", GroupErrorCode.NO_AUTHORIZATION_FOR_ITEM_EXCEPTION.getHttpStatus());
+        }
+
+        if (groupParticipation.getGroup().getHostId().equals(userId)) {
+            throw new BaseException("공고 작성자는 탈퇴를 할 수 없고 삭제만 가능합니다.", GroupErrorCode.ILLEGAL_REQUEST.getHttpStatus());
         }
 
         groupParticipationRepository.delete(groupParticipation);
